requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
accelerate>=0.20.0
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.30.3
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
filelock==3.29.0
fire==0.5.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
groundingdino-py==0.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.12
ImageIO==2.37.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.65.0
numpy==1.26.4
onnxruntime==1.23.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==26.1
Pillow==9.5.0
platformdirs==4.9.6
pooch==1.9.0
protobuf==7.34.1
pycocotools==2.0.11
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyMatting==1.1.15
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
rembg==2.0.69
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.15.3
segment-anything-py==1.0.1
shellingham==1.5.4
simple-lama-inpainting==0.1.2
six==1.17.0
sniffio==1.3.1
supervision==0.6.0
sympy==1.14.0
termcolor==3.3.0
tifffile==2025.5.10
timm==1.0.26
tokenizers==0.13.3
tomli==2.4.1
torch==2.11.0+cu128
torchvision==0.26.0+cu128
tqdm==4.67.3
transformers==4.30.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
yapf==0.43.0
